# Optimize AI update loop