Difference especially noticable with increased "exterior cell loading distance".
Affected lines take a most of mechanics update time.

Also this code is a quite strange. As I understand, it supposed to reset the HitAttemptActorId for player, if target actor disappeared from cell for some reason. But we do not set HitAttemptActorId inside affected loop, only reset it, so there is no point to run this code for every actor in scene - it will use the same value or -1.
So run this check only once before AI update.

I decided to do not create a separate task since we already have one for AI optimization in 0.45 milestone.

Also this PR optimizes combat music update a bit.

A little performance comparison: time in milliseconds to update game mechanics (based on profilier output), compared with master on i7-7700 in Vivec:

exterior cell load distance = 4
Master: 12.5ms
PR: 3 ms

exterior cell load distance = 3
Master: 6ms
PR: 2ms

exterior cell load distance = 2
Master: 2ms
PR: 1ms

exterior cell load distance = 1
Master: 0.8ms
PR: 0.6ms