@Capostrophic 2018-09-21
Boolean from isAIActive() is similarly recovered for each and every actor in updateCombatMusic() unnecessarily - and since the related code isn't even used if AI is disabled, the whole actor iterating loop there there could be put behind the check for the very same boolean. updateCombatMusic() is called every frame from the actor update loop.