@akortunov 2018-09-21
I rewrote the updateCombatMusic(), but there is not noticable performance gain since this function does not use heavy operations (such as search).