@Capostrophic 2018-09-21
Every bit helps.