@akortunov 2018-09-22
Added optimization for drowning state update: previously we used a search inside mActors map to retrieve actor's character controller just to check if actor is knocked out. Just pass this flag as parameter now.

I think I did what I could do.